Corretly map to Subject enum
